refactor: more restrucutre
